fix(emacs): include `htmlize` dependency for org-html-export
Problem: my blog used to reliably produce syntax-highlighted code blocks
following the syntax highlighting of this Emacs configuration. However,
recently, that stopped. This was a result of `htmlize` not being up to date:

> Warning (emacs): Cannot fontify source block (htmlize.el >= 1.34 required).
> Falling back to plain text.  (see `org-html-htmlize-output-type')

To resolve this error, I pull in the dependency explicitly, which should result
in org export using the correct, updated version.

Signed-off-by: Lucas Sta Maria <lucas@priime.dev>